pyproject.toml
[project]
name = "ai-images-classifier"
version = "0.1.0"
description = "AI Images Classifier with PyTorch Lightning and Hydra"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.4.2",
    "hydra-core>=1.3.2",
    "lightning>=2.6.0",
    "omegaconf>=2.3.0",
    "pillow>=12.1.0",
    "scikit-learn>=1.8.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.14.11",
    "tensorboard>=2.20.0",
]
